pyproject.toml
[tool.poetry]
name = "target-gcs"
version = "0.0.1"
description = "`target-gcs` is a Singer target for GCS, built with the Meltano SDK for Singer Targets."
authors = ["Datateer Ops"]
keywords = [
    "ELT",
    "GCS",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.11,>=3.7.1"
requests = "^2.25.1"
singer-sdk = "^0.5.0"
google-cloud-storage = "^2.3.0"
google-api-python-client = "^2.49.0"
smart-open = {extras = ["gcs"], version = "^6.0.0"}
pytest-watch = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"
black = "^22.3.0"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "target_gcs"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
target-gcs = 'target_gcs.target:TargetGCS.cli'



[tool.black]
py36 = true
include = '\.pyi?$'
exclude = '.meltano/'